refactor dialogs more
