+++ src/ui/dialogs/treeComponent.ts
@@ -1,6 +1,6 @@
 // [Smoosic](https://github.com/AaronDavidNewman/Smoosic)
 // Copyright (c) Aaron David Newman 2021.
-import { SuiComponentBase, SuiDialogNotifier } from '../dialogComponents';
+import { SuiComponentBase, SuiDialogNotifier } from './components/baseComponent';
 import { htmlHelpers } from '../../common/htmlHelpers';
 declare var $: any;
